==> 00-gnad-scraper.ipynb <==
# Global Nonviolent Action Database
- Scraping here: https://nvdatabase.swarthmore.edu/
- Get details for North American campaigns only (U.S., Mexico, Canada).

import pandas as pd
import numpy as np
import requests
import time
import unicodedata
import math
import regex as re
from bs4 import BeautifulSoup
from ast import literal_eval

## Functions

def get_country_slugs(country):
    '''
    Get slugs to get urls for individual campaigns
    '''
    url = f'{root_url}/browse/{country}/all/all/all/all'
    res = requests.get(url)
    soup = BeautifulSoup(res.content, 'lxml')

    # Get number of pages to go through from "Showing 1-{showing} of {total_results}"
    results = soup.find('div', {'class': 'view-browse-cases'}
                       ).find('div', {'class': 'view-header'}).text.strip()
    total_results = int(re.search(r'(?<= of )\d*?(?= results)', results)[0])
    showing = int(re.search(r'(?<=-)\d*?(?= of)', results)[0])
    pages = math.ceil(total_results/showing)

    articles = []

    counter = 0
    print('Getting slugs... Directory page # ')

    # Cycle through all pages in directory
    for page in range(pages):
        url = f'{root_url}/browse/{country}/all/all/all/all?page={page}'
        res = requests.get(url)
        soup = BeautifulSoup(res.content, 'lxml')

        for article in soup.find_all('article'):
            articles.append(article['about'])

        print(counter+1, end=' ')

        time.sleep(1)
        counter += 1
    
    return articles

def get_campaign_details(country_slugs):
    '''
    Get details for individual campaigns
    '''
    campaigns = []
    counter = 0
    total_campaigns = len(country_slugs)

    print(f'Getting {total_campaigns} campaign details... ')
    for article in range(total_campaigns):
        url = f'{root_url}{country_slugs[article]}'
        res = requests.get(url)
        soup = BeautifulSoup(res.content, 'lxml')

        campaign = {}
        
        campaign['title'] = soup.find('h1', {'class': 'page-header'}
                 ).find('span').text.strip()
        
        try:
            campaign['goals'] = soup.find('div', {'class': 'field--name-field-goals'}
                     ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['movementcluster'] = soup.find('div', {'class': 'field--name-field-movementcluster'}
                                       ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['time_period'] = soup.find('div', {'class': 'field--name-field-startime'}
                                   ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['country'] = soup.find('div', {'class': 'field--name-field-loc-country'}
                               ).find_all('div', {'class': 'field--item'})
            campaign['country'] = [x.text.strip() for x in campaign['country']]
        except:
            pass

        try:
            campaign['location'] = soup.find('div', {'class': 'field--name-field-city'}
                                ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['location_desc'] = soup.find('div', {'class': 'field--name-field-loc-remainder'}
                                     ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['pcs'] = soup.find('div', {'class': 'field--name-field-pcs-tags'}
                           ).find_all('a')
            campaign['pcs'] = [x.text.strip() for x in campaign['pcs']]
        except:
            pass

        methods = dict()
        for segment in range(1,7):
            try:
                methods[segment] = soup.find('div', {'class': f'field--name-field-tactic{segment}'}
                           ).find_all('a')
                methods[segment] = [method.text.strip() for method in methods[segment]]
            except:
                pass    
        try:
            methods['addl'] = soup.find('div', {'class': 'field--name-field-tactic7'}
                                ).find_all('a')
            methods['addl'] = [x.text.strip() for x in methods['addl']]
        except:
            pass
        campaign['methods'] = methods

        try:
            campaign['segment_length'] = soup.find('div', {'class': 'field--name-field-segmenth-length'}
                                      ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['leaders'] = soup.find('div', {'class': 'field--name-field-leaders'}
                               ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['partners'] = soup.find('div', {'class': 'field--name-field-partners'}
                            ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['allies'] = soup.find('div', {'class': 'field--name-field-external-allies'}
                              ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['social_elites'] = soup.find('div', {'class': 'field--name-field-involvelemt'}
                                 ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['opponents'] = soup.find('div', {'class': 'field--name-field-opponents'}
                                 ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pas

        try:
            campaign['opp_nvresponses'] = soup.find('div', {'class': 'field--name-field-nvresponses'}
                                       ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass
        
        try:
            campaign['campaigner_violence'] = soup.find('div', {'class': 'field--name-field-campaignviolence'}
                                           ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['repressive_violence'] = soup.find('div', {'class': 'field--name-field-repressviolence'}
                                           ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass
        
        try:
            campaign['cluster'] = soup.find('div', {'class': 'field--name-field-cluster'}
                               ).find_all('div', {'class': 'field--item'})
            campaign['cluster'] = [x.text.strip() for x in campaign['cluster']]
        except:
            pass

        try:
            campaign['classification'] = soup.find('div', {'class': 'field--name-field-classification'}
                                      ).find_all('div', {'class': 'field--item'})
            campaign['classification'] = [x.text.strip() for x in campaign['classification']]
        except:
            pass

        try:
            campaign['group'] = soup.find('div', {'class': 'field--name-field-groupchar'}
                             ).find_all('div', {'class': 'field--item'})
            campaign['group'] = [x.text.strip() for x in campaign['group']]
        except:
            pass

        groups = dict()
        for segment in range(1,7):
            try:
                groups[segment] = soup.find('div', {'class': f'field--name-field-group{segment}'}
                                           ).find_all('div', {'class': 'field--item'})
                groups[segment] = [group.text.strip() for group in groups[segment]]
            except:
                pass
        campaign['groups'] = groups

        try:
            campaign['success_goal'] = soup.find('div', {'class': 'field--name-field-procedure'}
                                    ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['success_survival'] = soup.find('div', {'class': 'field--name-field-survivalgoals'}
                                        ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['success_growth'] = soup.find('div', {'class': 'field--name-field-growth'}
                                      ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass

        try:
            campaign['success_total'] = soup.find('div', {'class': 'field--name-field-total-points'}
                                     ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass
        
        try:
            campaign['success_notes'] = soup.find('div', {'class': 'field--name-field-notesoutcome'}
                                     ).find('div', {'class': 'field--item'}).text.strip()
        except:
            pass
        
        campaign['url'] = url
        
        campaigns.append(campaign)
        
        if (counter+1) % 5 == 0:
            print(counter+1, end=' ')

        time.sleep(1)
        counter += 1
        if counter == total_campaigns:
            print('Done.')

    # Save to dataframe
    campaigns = pd.DataFrame(campaigns)
    
    return campaigns

def dict_to_cols(df, col):
    '''
    Use to change 'methods' and 'groups' from dict to multiple columns
    https://stackoverflow.com/questions/38231591/split-explode-a-column-of-dictionaries-into-separate-columns-with-pandas
    '''
    # replace NaN with '{}' if the column is strings, otherwise replace with {}
    df[col] = df[col].fillna({i: {} for i in df.index})  # if the column is not strings

    # normalize the column of dictionaries and join it to df
    df = df.join(pd.json_normalize(df[col]).add_prefix(f'{col}_'))

    # drop col
    df.drop(columns=[col], inplace=True)
    
    return df

def add_missing_cols(df):
    '''
    Add any missing columns and reorder.
    '''
    cols = ['title', 'goals', 'movementcluster', 'time_period', 'country', 'location', 'location_desc', 'pcs',
            'methods_1', 'methods_2', 'methods_3', 'methods_4', 'methods_5', 'methods_6', 'methods_addl',
            'segment_length', 'leaders', 'partners', 'allies', 'social_elites',
            'opponents', 'opp_nvresponses', 'campaigner_violence', 'repressive_violence',
            'cluster', 'classification', 'group',
            'groups_1', 'groups_2', 'groups_3', 'groups_4', 'groups_5', 'groups_6',
            'success_goal', 'success_survival', 'success_growth', 'success_total', 'success_notes', 'url']
    for col in cols:
        if col not in df:
            df[col] = np.nan
            
    df = df[cols]
    return df

def campaigns(country_slugs):
    df = get_campaign_details(country_slugs)
    df = dict_to_cols(df, 'methods')
    df = dict_to_cols(df, 'groups')
    df = add_missing_cols(df)
    return df

## Get all campaign info

root_url = 'https://nvdatabase.swarthmore.edu'

us_slugs = get_country_slugs('United%20States')
mexico_slugs = get_country_slugs('Mexico')
canada_slugs = get_country_slugs('Canada')

us = campaigns(us_slugs)
mexico = campaigns(mexico_slugs)
canada = campaigns(canada_slugs)

## Combine and save to .csv

all_campaigns = pd.concat([us, mexico, canada])

# Save to .csv
all_campaigns.to_csv('./campaigns.csv', index=False)

## Preview

all_campaigns

Next steps:
- Could get start and end date from `time_period`. Depending on earliest / lastest, can bin by decade, presidency, etc.
- `pcs`: Create dummies AND/OR count of different tags
- `cluster`: count or dummies
- `methods`: Could create just a couple dummies - strike, boycott, sit-in...
- `methods` count: count of total different methods used
- `partners/ allies/ elites`: either dummy or count
- opponents / violence: dummy yes or no
- `success`: Change just to first digits.
- others - get count of number of different groups...
